requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python-headless
h5py
python-multipart
transformers
pillow
torch